backend/auth_service/requirements.txt
python-json-logger==2.0.7
aiobotocore==2.13.1
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aioitertools==0.11.0
aiosignal==1.3.1
aiosmtplib==2.0.2
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asyncpg==0.29.0
attrs==24.2.0
bcrypt==4.1.2
bidict==0.23.1
blinker==1.8.2
botocore==1.34.131
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
dnspython==2.6.1
email_validator==2.1.2
fastapi==0.112.2
fastapi-cache2==0.2.2
fastapi-cli==0.0.4
fastapi-mail==1.4.1
fastapi-users==13.0.0
fastapi-users-db-sqlalchemy==6.0.1
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib_metadata==8.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
makefun==1.15.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
orjson==3.10.6
packaging==24.1
passlib==1.7.4
pendulum==3.0.0
psycopg2==2.9.9
pwdlib==0.2.0
pycparser==2.22
pycron==3.0.0
pydantic==2.8.2
pydantic-extra-types==2.9.0
pydantic-settings==2.3.4
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.3
pytz==2024.1
PyYAML==6.0.1
redis==5.0.8
requests==2.32.3
rich==13.7.1
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
SQLAlchemy-Utils==0.41.2
sqlmodel==0.0.20
starlette==0.37.2
taskiq==0.11.7
taskiq-dependencies==1.5.3
taskiq-redis==1.0.0
time-machine==2.15.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
watchfiles==0.22.0
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.20.1
pytest==8.3.3
asyncpg
sqlalchemy
pydantic_settings
httpx